+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Change Log
 ==========
 
-Unreleased
-----------
+v1.0.4
+------
 - Added `timeout` parameter to database initializer (SUHAR1K)
 - Added `verify_ssl_cert` parameter to database initializer
 - Added `final()` method to querysets (M1hacka)